+++ nengo/synapses.py
@@ -112,6 +112,25 @@ class LinearFilter(Synapse):
         return "%s(%s, %s, analog=%r)" % (
             self.__class__.__name__, self.num, self.den, self.analog)
 
+    def evaluate(self, frequencies):
+        """Evaluate transfer function at given frequencies.
+
+        Example
+        -------
+        Using the ``evaluate`` function to make a Bode plot::
+        >>> synapse = nengo.synapses.LinearFilter([1], [0.02, 1])
+        >>> f = numpy.logspace(-1, 3, 100)
+        >>> y = synapse.evaluate(f)
+        >>> plt.subplot(211); plt.semilogx(f, 20*np.log10(np.abs(y)))
+        >>> plt.xlabel('frequency [Hz]'); plt.ylabel('magnitude [dB]')
+        >>> plt.subplot(212); plt.semilogx(f, np.angle(y))
+        >>> plt.xlabel('frequency [Hz]'); plt.ylabel('phase [radians]')
+        """
+        frequencies = 2.j*np.pi*frequencies
+        w = frequencies if self.analog else np.exp(frequencies)
+        y = np.polyval(self.num, w) / np.polyval(self.den, w)
+        return y
+
     def make_step(self, dt, output, method='zoh'):
         num, den = self.num, self.den
         if self.analog: